增加错误处理方法

+++ httpo/response.go
@@ -1,7 +1,7 @@
 package httpo
 
 type Response struct {
-	Code int64  `json:"code"`
+	Code int    `json:"code"`
 	Msg  string `json:"msg"`
 	Data any    `json:"data"`
 }
